+++ Datatypes/DtAsignatura.h
@@ -15,19 +15,19 @@ private:
     bool teorico;
     bool practico;
     bool monitoreo;
-    float tiempoTotalDictado;
+    double tiempoTotalDictado;
 public:
     DtAsignatura();
     ~DtAsignatura();
-    DtAsignatura(string codigoAsignatura, string nombreAsignatura, bool teorico, bool practico, bool monitoreo, float tiempoTotalDictado);
+    DtAsignatura(string codigoAsignatura, string nombreAsignatura, bool teorico, bool practico, bool monitoreo, double tiempoTotalDictado);
     DtAsignatura(string codigoAsignatura, string nombreAsignatura, bool teorico, bool practico, bool monitoreo);
-    DtAsignatura(string nombreAsignatura, float tiempoTotalDictado);
+    DtAsignatura(string nombreAsignatura, double tiempoTotalDictado);
     string getCodigoAsignatura() const;
     string getNombreAsignatura() const;
     bool tieneTeorico() const;
     bool tienePractico() const;
     bool tieneMonitoreo() const;
-    float getTiempoTotalDictado() const;
+    double getTiempoTotalDictado() const;
     void mostrarInfo(ostream&);
     friend ostream& operator<<(ostream&, DtAsignatura*);
 };
